Split load_from_drive into client & server files.
